requirements.txt
arrow
uvicorn[standard]
fastapi==0.111.0
python-dotenv==1.0.1
pydantic-settings==2.3.4
databases==0.9.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
httpx==0.27.0
email-validator==2.2.0
